libft_v0.2.1

+++ ft_strrchr.c
@@ -6,7 +6,7 @@
 /*   By: rsoto-in <rsoto-in@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/15 15:57:53 by rsoto-in          #+#    #+#             */
-/*   Updated: 2019/12/09 12:45:43 by rsoto-in         ###   ########.fr       */
+/*   Updated: 2020/01/11 12:48:33 by rsoto-in         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strrchr(const char *s, int c)
 {
-	size_t len;
+	int	len;
 
 	len = ft_strlen(s);
 	if (!c && s[len] == c)
